KEYLOGGER
This is the file that holds the keylogger that you can use on your own system.

-IMPORTANT-
NONE OF THE DATA TAKEN FROM THIS PROGRAM GOES ANYWHERE BUT YOUR DEVICE. All the data logged within this program stays natively in your OWN document folder on your system, only you are able to see the secret.txt text file created on your system, and the data logged is just for educational purposes only. If you do wish to run this program on your system, one, I suggest running this on a VM (Virtual Machine) if you are worried about your system, and two, you will need to disable your antivirus software like Windows Defender and accept all the warnings from your browser when downloading the file. For Windows Defender, you will only need to disable one setting, Real-time protection. Go to Windows Security > Virus & Threat protection Settings > Real-time protection > off. To terminate the file, just go to your task manager and look for the name for the name of the program or "keylogger.exe" and terminate the process. This ends the process and you can delete the secret.txt text file created in your system. While this malware gets picked up by most antivirus softwares, there are a good number of malwares out there that don't, so remember to always be safe on the internet and to never download suspicious links.